#1934 develop global setting management plugin - BABAK (#1958)
* global setting update is added

* superadmin permission is applied

* setup manager is removed from permission manager

* some code cleanup

* Refactor GlobalSettings handling and update API responses

Enhanced validation and exception handling for the SuperAdmins list in GlobalSettingsService. Updated SetupManager to initialize SuperAdmins correctly and removed unused InitLayouts method. Refactored UserService to improve exception messages and added cancellationToken parameter.

Changed GlobalSettingsController methods to return GlobalSettingsResponse and updated mapping configurations accordingly. Updated GeneratedApiClients to reflect changes in response types and serialization properties.

Refactored GlobalSettingsPlugin to use a new GlobalSettingsUpdateModel class for form data management. Updated form binding, initialization, and submission logic. Added and renamed exception codes for clarity.

---------

+++ src/Backend/FluentCMS.Services/UserService.cs
@@ -107,10 +107,16 @@ public class UserService(
         var user = await userManager.FindByIdAsync(id.ToString())
             ?? throw new AppException(ExceptionCodes.UserNotFound);
 
+        var globalSettings = await globalSettingsRepository.Get(cancellationToken) ??
+            throw new AppException(ExceptionCodes.GlobalSettingsNotFound);
+
+        if (globalSettings.SuperAdmins.Contains(user.UserName))
+            throw new AppException(ExceptionCodes.UserSuperAdminCanNotBeDeleted);
+
         var userRemoveResult = await userManager.DeleteAsync(user);
         userRemoveResult.ThrowIfInvalid();
 
-        await DeleteUserAllRoles(user.Id);
+        await DeleteUserAllRoles(user.Id, cancellationToken);
 
         return true;
     }
